[BT-66] move setup to setup/index.ts, extract firebase-admin mock, clean up setup, update mock, add google-signin tests
- Moved e2e test setup from 'setup.e2e.ts' to 'setup/index.ts'
- Moved firebase-admin mock definition to '__mocks__/firebase-admin.ts'
- Removed unnecessary code from test setup
- Updated firebase-admin mock to return user email in verifyIdToken
- Added tests for 'google-signin' route in auth tests
